refactor links to file, update README
